--- a/remote_write.go
+++ b/remote_write.go
@@ -2,6 +2,7 @@ package remotewrite
 
 import (
 	"bytes"
+	"encoding/binary"
 	"fmt"
 	"log"
 	"math"
@@ -23,6 +24,7 @@ import (
 	"go.k6.io/k6/js/modules"
 	"go.k6.io/k6/lib"
 	"go.k6.io/k6/lib/netext/httpext"
+	"google.golang.org/protobuf/encoding/protowire"
 )
 
 // Register the extension on module initialization, available to
@@ -49,9 +51,10 @@ func (r *remoteWriteModule) NewModuleInstance(vu modules.VU) modules.Instance {
 func (r *RemoteWrite) Exports() modules.Exports {
 	return modules.Exports{
 		Named: map[string]interface{}{
-			"Client":     r.xclient,
-			"Sample":     r.sample,
-			"Timeseries": r.timeseries,
+			"Client":                   r.xclient,
+			"Sample":                   r.sample,
+			"Timeseries":               r.timeseries,
+			"precompileLabelTemplates": compileLabelTemplates,
 		},
 	}
 }
@@ -325,9 +328,16 @@ func compileTemplate(template string) (*labelGenerator, error) {
 	}
 	switch template[i+len("${series_id")] {
 	case '}':
-		return &labelGenerator{ToString: func(seriesID int) string {
-			return template[:i] + strconv.Itoa(seriesID) + template[i+len("${series_id}"):]
-		}}, nil
+		return &labelGenerator{
+			ToString: func(seriesID int) string {
+				return template[:i] + strconv.Itoa(seriesID) + template[i+len("${series_id}"):]
+			},
+			AppendByte: func(b []byte, seriesID int) []byte {
+				b = append(b, template[:i]...)
+				b = strconv.AppendInt(b, int64(seriesID), 10)
+				return append(b, template[i+len("${series_id}"):]...)
+			},
+		}, nil
 	case '%':
 		end := strings.Index(template[i:], "}")
 		if end == -1 {
@@ -337,14 +347,28 @@ func compileTemplate(template string) (*labelGenerator, error) {
 		if err != nil {
 			return nil, fmt.Errorf("can't parse divisor of the module operator %w", err)
 		}
+
+		possibleValues := make([][]byte, d)
+		// TODO have an upper limit
+		for j := 0; j < d; j++ {
+			var b []byte
+			b = append(b, template[:i]...)
+			b = strconv.AppendInt(b, int64(j), 10)
+			possibleValues[j] = append(b, template[i+end+1:]...)
+		}
 		possibleValuesS := make([]string, d)
 		// TODO have an upper limit
 		for j := 0; j < d; j++ {
 			possibleValuesS[j] = template[:i] + strconv.Itoa((j)) + template[i+end+1:]
 		}
-		return &labelGenerator{ToString: func(seriesID int) string {
-			return possibleValuesS[seriesID%d]
-		}}, nil
+		return &labelGenerator{
+			ToString: func(seriesID int) string {
+				return possibleValuesS[seriesID%d]
+			},
+			AppendByte: func(b []byte, seriesID int) []byte {
+				return append(b, possibleValues[seriesID%d]...)
+			},
+		}, nil
 	case '/':
 		end := strings.Index(template[i:], "}")
 		if end == -1 {
@@ -357,6 +381,8 @@ func compileTemplate(template string) (*labelGenerator, error) {
 		var memoizeS string
 		var memoizeSValue int
 
+		var memoize []byte
+		var memoizeValue int64
 		return &labelGenerator{
 			ToString: func(seriesID int) string {
 				value := (seriesID / d)
@@ -366,18 +392,31 @@ func compileTemplate(template string) (*labelGenerator, error) {
 				}
 				return memoizeS
 			},
+			AppendByte: func(b []byte, seriesID int) []byte {
+				value := int64(seriesID / d)
+				if memoize == nil || value != memoizeValue {
+					memoizeValue = value
+					memoize = memoize[:0]
+					memoize = append(memoize, template[:i]...)
+					memoize = strconv.AppendInt(memoize, value, 10)
+					memoize = append(memoize, template[i+end+1:]...)
+				}
+				return append(b, memoize...)
+			},
 		}, nil
 	}
 	return nil, errors.New("unsupported template")
 }
 
 type labelGenerator struct {
-	ToString func(int) string
+	ToString   func(int) string
+	AppendByte func([]byte, int) []byte
 }
 
 func newIdentityLabelGenerator(t string) *labelGenerator {
 	return &labelGenerator{
-		ToString: func(int) string { return t },
+		ToString:   func(int) string { return t },
+		AppendByte: func(b []byte, _ int) []byte { return append(b, t...) },
 	}
 }
 
@@ -395,7 +434,7 @@ func generateFromTemplates(r *rand.Rand, minValue, maxValue int,
 	for seriesID := minSeriesID; seriesID < maxSeriesID; seriesID++ {
 		labels := make([]prompb.Label, len(labelsTemplate))
 		// TODO optimize
-		for i, template := range compiledTemplates {
+		for i, template := range compiledTemplates.compiledTemplates {
 			labels[i] = prompb.Label{Name: template.name, Value: template.generator.ToString(seriesID)}
 		}
 
@@ -413,12 +452,17 @@ func generateFromTemplates(r *rand.Rand, minValue, maxValue int,
 	return series, nil
 }
 
+// this is opaque on purpose so that it can't be done anything to from the js side
+type labelTemplates struct {
+	compiledTemplates []compiledTemplate
+	labelValue        []byte
+}
 type compiledTemplate struct {
 	name      string
 	generator *labelGenerator
 }
 
-func compileLabelTemplates(labelsTemplate map[string]string) ([]compiledTemplate, error) {
+func compileLabelTemplates(labelsTemplate map[string]string) (*labelTemplates, error) {
 	compiledTemplates := make([]compiledTemplate, len(labelsTemplate))
 	{
 		i := 0
@@ -435,7 +479,10 @@ func compileLabelTemplates(labelsTemplate map[string]string) ([]compiledTemplate
 	sort.Slice(compiledTemplates, func(i, j int) bool {
 		return compiledTemplates[i].name < compiledTemplates[j].name
 	})
-	return compiledTemplates, nil
+	return &labelTemplates{
+		compiledTemplates: compiledTemplates,
+		labelValue:        make([]byte, 128), // this is way more than necessary and it will grow if needed
+	}, nil
 }
 
 func (c *Client) StoreFromTemplates(
@@ -449,6 +496,98 @@ func (c *Client) StoreFromTemplates(
 		return httpext.Response{}, err
 	}
 	return c.store(ts)
+}
+
+func (template *labelTemplates) writeFor(w *bytes.Buffer, value float64, seriesID int, timestamp int64) (err error) {
+	labelValue := template.labelValue[:]
+	for _, template := range template.compiledTemplates {
+		labelValue = labelValue[:0]
+		w.WriteByte(0xa)
+		labelValue = protowire.AppendVarint(labelValue, uint64(len(template.name)))
+		n1 := len(labelValue)
+		labelValue = template.generator.AppendByte(labelValue, seriesID)
+		n2 := len(labelValue)
+		labelValue = protowire.AppendVarint(labelValue, uint64(n2-n1))
+		n3 := len(labelValue)
+
+		labelValue = protowire.AppendVarint(labelValue, uint64(n3+1+1+len(template.name)))
+		w.Write(labelValue[n3:])
+		w.WriteByte(0xa)
+		w.Write(labelValue[:n1])
+		w.WriteString(template.name)
+		w.WriteByte(0x12)
+		w.Write(labelValue[n2:n3])
+		w.Write(labelValue[n1:n2])
+	}
+
+	labelValue = labelValue[:10]
+	labelValue[0] = 0x9
+	binary.LittleEndian.PutUint64(labelValue[1:9], uint64(math.Float64bits(value)))
+	labelValue[9] = 0x10
+	labelValue = protowire.AppendVarint(labelValue, uint64(timestamp))
+
+	n := len(labelValue)
+	labelValue = labelValue[:n+1]
+	labelValue[n] = 0x12
+	labelValue = protowire.AppendVarint(labelValue, uint64(n))
+	w.Write(labelValue[n:])
+	w.Write(labelValue[:n])
+	template.labelValue = labelValue
+	return nil // TODO fix
+}
+
+func (c *Client) StoreFromPrecompiledTemplates(
+	minValue, maxValue int,
+	timestamp int64, minSeriesID, maxSeriesID int,
+	template *labelTemplates,
+) (httpext.Response, error) {
+	state := c.vu.State()
+	if state == nil {
+		return *httpext.NewResponse(), errors.New("State is nil")
+	}
+	r := rand.New(rand.NewSource(time.Now().Unix()))
+	buf := generateFromPrecompiledTemplates(r, minValue, maxValue, timestamp, minSeriesID, maxSeriesID, template)
+	b := buf.Bytes()
+	compressed := make([]byte, len(b)/9) // the general size is actually between 1/9 and 1/10th but this is closed enough
+	compressed = snappy.Encode(compressed, b)
+
+	res, err := c.send(state, compressed)
+	if err != nil {
+		return *httpext.NewResponse(), errors.Wrap(err, "remote-write request failed")
+	}
+	res.Request.Body = ""
+
+	return res, nil
+}
+
+func generateFromPrecompiledTemplates(
+	r *rand.Rand,
+	minValue, maxValue int,
+	timestamp int64, minSeriesID, maxSeriesID int,
+	template *labelTemplates,
+) *bytes.Buffer {
+	bigB := make([]byte, 1024)
+	buf := new(bytes.Buffer)
+	buf.Reset()
+
+	tsBuf := new(bytes.Buffer)
+	bigB[0] = 0xa
+	template.writeFor(tsBuf, valueBetween(r, minValue, maxValue), minSeriesID, timestamp)
+	bigB = protowire.AppendVarint(bigB[:1], uint64(tsBuf.Len()))
+	buf.Write(bigB)
+	tsBuf.WriteTo(buf)
+
+	buf.Grow((buf.Len() + 2) * (maxSeriesID - minSeriesID)) // heuristics to try to get big enough buffer in one go
+	for seriesID := minSeriesID + 1; seriesID < maxSeriesID; seriesID++ {
+		tsBuf.Reset()
+		bigB[0] = 0xa
+		template.writeFor(tsBuf, valueBetween(r, minValue, maxValue), seriesID, timestamp)
+		bigB = protowire.AppendVarint(bigB[:1], uint64(tsBuf.Len()))
+		buf.Write(bigB)
+		tsBuf.WriteTo(buf)
+	}
+
+	return buf
 }
 
 func valueBetween(r *rand.Rand, min, max int) float64 {