Benchmarks results comparing first and last:
```
name                        old time/op    new time/op    delta
CompileTemplatesSimple-8       106ns ± 7%     984ns ±10%  +825.96%  (p=0.000 n=10+10)
CompileTemplatesComplex-8      181ns ± 8%    1415ns ± 6%  +681.53%  (p=0.000 n=10+10)
EvaluateTemplatesSimple-8      206ns ±10%      25ns ± 8%   -88.07%  (p=0.000 n=10+10)
EvaluateTemplatesComplex-8     186ns ± 7%       7ns ±10%   -96.24%  (p=0.000 n=10+10)
StoreFromTemplates-8           205ms ±10%     158ms ±10%   -23.00%  (p=0.000 n=10+9)

name                        old alloc/op   new alloc/op   delta
StoreFromTemplates-8           157MB ± 0%     117MB ± 0%   -25.35%  (p=0.000 n=9+10)

name                        old allocs/op  new allocs/op  delta
StoreFromTemplates-8            589k ± 0%        0k ± 1%   -99.95%  (p=0.000 n=10+10)
```

The really important part is that it no longer near close to 600k allocations but instead some three-digit number. This significantly reduces the CPU usage in the real world. 

If we compare
```
GenerateFromTemplatesAndMarshal-8  84.2ms ±15%
```
from the first commit - this is the part that has been heavily optimized and in the end replaced to
```
GenerateFromPrecompiledTemplates-8  27.4ms ±20%
```
(which does include the marshalling actually) you can see almost 4x improvement.

From some experiments it seems that a real world improvement will be in that vicinity if not better actually. 

This now also works for the "old" API. So we might think about not actually adding JS precompilation call and then using that precompilation if it's not needed :thinking: 

Additionally, template parsing errors are now imported back instead of being ignored and just using the template as the value.