@MStoykov 2022-03-23
This is #15 but in actually understandable commits