@MStoykov 2022-03-23
>I don't think it's a high risk change

I mean it changes [how we marshal to protobuf](https://github.com/grafana/xk6-client-prometheus-remote/pull/17/files#diff-fec32fe45234f16d1e69465b42b55e1ef20b80c6facdb06a63bb989347f3f409R444-R479) in order to get the allocations down so low.

I can only hope that between all the test we've run and https://github.com/grafana/xk6-client-prometheus-remote/pull/17/files#diff-c7b61c54e01102d2023bb4d5bf118795021d724bbef1f4ea661ec86e3f59d2fdR175 it will be fine.


 I will wait for one more :+1: as we are not in a hurry to merge this